reorganized source files & improved code
